(fix(repo-automation)): split command registration helpers to close blocker
- Extract admin/support and feature-workflow registrations into focused modules while preserving command IDs and the public registration entrypoint
- Refresh remediation evidence and review artifacts, including archiving the earlier audit set under a timestamped audit folder
- Record a clean TypeScript QA pass and explicit development-based PR context for the residual blocker closeout

Refs: #164
