+++ docs/features/active/2026-04-26-codex-native-converter-164/code-review.2026-04-26T20-05.md
@@ -1,0 +1,23 @@
+# Code Review: codex-native-converter remediation closeout (#164)
+
+**Reviewer:** atomic_executor (GitHub Copilot)
+**Scope:** Post-remediation branch state for `feature/codex-native-converter-164` against explicit base `development`
+**Base commit:** `0762f58a1451994999c2f49f2dbdc489120d138a`
+**Head commit:** `b9542764a8271b83ecb075b7ca6edeb8575d1dfe`
+
+## Executive Summary
+
+The remediation is complete for the residual command-registration blocker. The repo-automation registration surface is now split into focused admin/support and feature-workflow helper modules, while `repo-automation-command-registration.ts` remains as a thin public assembly layer. Existing command IDs, prompt flows, and the `extension.ts` activation wiring were preserved, and the clean TypeScript QA pass completed successfully after a formatter restart.
+
+## Findings Table
+
+| Severity | File | Location | Finding | Recommendation | Rationale | Evidence |
+| --- | --- | --- | --- | --- | --- | --- |
+| Info | `extensions/drm-copilot/src/repo-automation-command-registration.ts` | file scope | The coordinator file is now a 23-line assembly layer that delegates to focused helper modules. | Keep future repo-automation registrations grouped by cohesive command family instead of expanding the coordinator again. | The residual policy blocker is closed and the public registration surface remains straightforward to review. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-repo-automation-command-registration-lines-after.2026-04-26T19-48.md` |
+| Info | `extensions/drm-copilot/src/repo-automation-command-registration-admin.ts` | file scope | Admin and review-support command registrations now live in a dedicated helper module. | Preserve this boundary for collect-context, push-down, converter, sync, and tool-list commands. | The split keeps administrative workflows separate from feature-entry flows and improves cohesion. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-2-command-registration-split-boundary.2026-04-26T19-48.md` |
+| Info | `extensions/drm-copilot/src/repo-automation-command-registration-feature-workflows.ts` | file scope | Feature-promotion and issue-linking registrations now live in a dedicated helper module with 100% line coverage. | Preserve this boundary for short-name, issue-number, and work-mode prompt flows. | The module now groups the most prompt-heavy workflow commands in one focused location without changing behavior. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md` |
+| Info | `extensions/drm-copilot/src/extension.ts` | import and registration call sites | The activation entrypoint still imports and invokes `registerRepoAutomationCommands` from the same public module. | Keep `extension.ts` as a thin activation coordinator. | The remediation preserved extension activation behavior while changing only internal module boundaries. | `extensions/drm-copilot/src/extension.ts:26`, `extensions/drm-copilot/src/extension.ts:99` |
+
+## Verdict
+
+Ready for re-review. No residual blocker remains in the remediation scope, and the refreshed `development`-based review evidence is complete.
