+++ docs/features/active/2026-04-26-codex-native-converter-164/code-review.2026-04-26T20-12.md
@@ -1,0 +1,27 @@
+# Code Review: codex-native-converter final post-remediation review (#164)
+
+**Reviewer:** GitHub Copilot
+**Scope:** Current working tree for `feature/codex-native-converter-164` relative to explicit base `development`
+**Base commit:** `0762f58a1451994999c2f49f2dbdc489120d138a`
+**Head commit reference:** `b9542764a8271b83ecb075b7ca6edeb8575d1dfe`
+**Working-tree scope note:** This review includes the current uncommitted repo-automation command-registration split verified directly in the workspace.
+
+## Executive Summary
+
+The branch is ready from a code-review perspective. The Python converter implementation remains strongly typed, cohesive, and supported by clean Black, Ruff, Pyright, and Pytest evidence. The TypeScript integration layer is now structurally compliant after the final repo-automation command-registration split: the coordinator file is thin, helper modules are focused by responsibility, and `extension.ts` remains an activation coordinator instead of an accumulation point.
+
+No blocker-level findings remain in the current working tree.
+
+## Findings Table
+
+| Severity | File | Location | Finding | Recommendation | Rationale | Evidence |
+| --- | --- | --- | --- | --- | --- | --- |
+| Info | `scripts/dev_tools/codex_native_converter/*.py` | package scope | The authoritative converter remains in Python and stays strongly typed and Pyright-clean after remediation work on the TypeScript wrapper. | Keep new conversion rules and validation behavior in the Python package first, and continue using TypeScript as a thin command/MCP boundary. | This preserves the architecture documented in `spec.md` and avoids logic drift between the Python and TypeScript surfaces. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/final-python-typecheck.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/final-python-coverage-delta.md` |
+| Info | `extensions/drm-copilot/src/repo-automation-command-registration.ts` | file scope | The coordinator is now a thin public assembly layer at 19 lines in the current working tree. | Keep future repo-automation registrations grouped in helper modules rather than growing the coordinator again. | The residual structural blocker is closed and the public registration surface is now straightforward to review. | Direct working-tree inspection; `artifacts/pr_context.appendix.txt`; `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-repo-automation-command-registration-lines-after.2026-04-26T19-48.md` |
+| Info | `extensions/drm-copilot/src/repo-automation-command-registration-admin.ts` | file scope | Admin and repository-support registrations now live in a dedicated helper module at 254 lines. | Preserve this boundary for collect-context, push-down, converter, sync, and tool-list commands. | The split improves cohesion and keeps the non-feature workflow commands separate from feature-promotion prompts. | Direct working-tree inspection; `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-2-command-registration-split-boundary.2026-04-26T19-48.md` |
+| Info | `extensions/drm-copilot/src/repo-automation-command-registration-feature-workflows.ts` | file scope | Feature-entry and promotion registrations now live in a dedicated helper module at 260 lines with full line coverage in the final remediation evidence. | Preserve this boundary for short-name, issue-number, and work-mode prompt flows. | The module groups the prompt-heavy workflow commands in one focused location without changing behavior. | Direct working-tree inspection; `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md` |
+| Info | `extensions/drm-copilot/src/extension.ts` | import and registration call sites | The extension activation entrypoint remains a thin coordinator at 266 lines and still imports `registerRepoAutomationCommands` as the public integration surface. | Keep new repo-automation wiring behind focused helper modules so `activate()` remains easy to review. | The final split preserves behavior while restoring structural compliance. | Direct working-tree inspection; `extensions/drm-copilot/src/extension.ts`; `artifacts/pr_context.appendix.txt` |
+
+## Verdict
+
+Ready for merge review. The current branch state has no remaining code-review blockers, and no additional remediation loop is required.
