+++ docs/features/active/2026-04-26-codex-native-converter-164/code-review.2026-04-26T23-41.md
@@ -5,16 +5,17 @@
 
 ## Executive Summary
 
-The remediation resolved the prior structural blocker. `extension.ts` and `repo-automation-service.ts` are now below the repository 500-line production-file limit, the extracted helper modules preserve the existing command and service contracts, and the final TypeScript QA loop passed cleanly. The refreshed PR-context summary now points to an explicit commit range against `development`, so the rerun review scope is unambiguous.
+The remediation resolved the two original structural blockers on `extension.ts` and `repo-automation-service.ts`, and the final TypeScript QA loop passed cleanly. However, the extracted helper `repo-automation-command-registration.ts` is 513 lines, so the branch still carries one structural policy violation and is not yet ready for final re-review.
 
 ## Findings Table
 
 | Severity | File | Location | Finding | Recommendation | Rationale | Evidence |
 | --- | --- | --- | --- | --- | --- | --- |
+| Major | `extensions/drm-copilot/src/repo-automation-command-registration.ts` | file scope | The extracted helper file is 513 lines, which exceeds the repository 500-line production-file limit. | Split the registration helpers into smaller modules, for example by grouping PR-context, promotion, and converter command registrations separately. | The remediation moved code out of `extension.ts`, but it did not fully clear the underlying structural policy risk because the new production module is still oversized. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-repo-automation-command-registration-lines-after.2026-04-26T19-20.md` |
 | Info | `extensions/drm-copilot/src/extension.ts` | file scope | The remediation reduced the file from 687 lines to 268 lines by moving repo-automation registrations into `repo-automation-command-registration.ts`. | Keep new command registrations in focused helper modules so `activate()` remains a thin coordinator. | The previous file-size blocker is closed and the new structure is easier to maintain. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-extension-lines.2026-04-26T19-20.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-extension-lines-after.2026-04-26T19-20.md` |
 | Info | `extensions/drm-copilot/src/repo-automation-service.ts` | file scope | The remediation reduced the file from 560 lines to 473 lines by moving workflow option-building logic into `repo-automation-service-workflows.ts`. | Keep future workflow option assembly in helper modules so the service file remains within the repository budget. | The public service contract stayed stable while the policy violation was removed. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-repo-automation-service-lines.2026-04-26T19-20.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-repo-automation-service-lines-after.2026-04-26T19-20.md` |
 | Info | `artifacts/pr_context.summary.txt` | review scope metadata | The rerun PR context now uses a non-empty commit range against `origin/development`. | Preserve the refreshed summary and appendix with the remediation evidence for final branch review. | The review basis is explicit and no longer depends on an empty-range fallback note. | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-pr-context-status.2026-04-26T19-20.md`, `artifacts/pr_context.summary.txt` |
 
 ## Verdict
 
-Ready for re-review. No remaining blocker was identified in the remediation scope, and the final TypeScript QA loop passed with the structural policy issue closed.
+Not ready for final re-review. The original blockers were reduced, but `repo-automation-command-registration.ts` remains over the repository file-size limit and requires one more structural split.
