+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/phase0-instructions-read.2026-04-26T19-48.md
@@ -1,0 +1,9 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Policy Order: 1-5
+
+Ordered Files Read:
+1. `.github/copilot-instructions.md`
+2. `.github/instructions/general-code-change.instructions.md`
+3. `.github/instructions/general-unit-test.instructions.md`
+4. `.github/instructions/typescript-code-change.instructions.md`
+5. `.github/instructions/typescript-unit-test.instructions.md`
