+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-2-typescript-format.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: npm --prefix extensions/drm-copilot run format
+EXIT_CODE: 0
+Output Summary: Prettier ran successfully; all matched extension package source, test, and config files were reported unchanged.
