+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-2-typescript-typecheck.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: npm --prefix extensions/drm-copilot run typecheck
+EXIT_CODE: 0
+Output Summary: TypeScript compilation completed successfully for the extension package with no type errors.
