+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-repo-automation-command-registration-lines.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: pwsh -NoProfile -Command "(Get-Content 'extensions/drm-copilot/src/repo-automation-command-registration.ts' | Measure-Object -Line).Lines"
+EXIT_CODE: 0
+Output Summary: 513
