+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-final-verdict.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: document-remediation-2-final-verdict
+EXIT_CODE: 0
+Output Summary: Residual blocker closed. `extensions/drm-copilot/src/repo-automation-command-registration.ts` is now 23 lines, the TypeScript QA loop completed with a clean final pass after 2 formatter iterations, refreshed review artifacts `policy-audit.2026-04-26T20-05.md`, `code-review.2026-04-26T20-05.md`, and `feature-audit.2026-04-26T20-05.md` all validated successfully, and no additional remediation loop is required before re-review.
