+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-pr-context-status.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: poetry run python -m scripts.dev_tools.pr_context.collector --base development
+EXIT_CODE: 0
+Output Summary: PR context refreshed against explicit base `development`. The rerun used a commit-range review (`0762f58a1451994999c2f49f2dbdc489120d138a..b9542764a8271b83ecb075b7ca6edeb8575d1dfe`) rather than a working-tree appendix. Resolved base: `origin/development @ 0762f58a1451994999c2f49f2dbdc489120d138a`. Resolved head: `feature/codex-native-converter-164 @ b9542764a8271b83ecb075b7ca6edeb8575d1dfe`.
