+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-format.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: npm --prefix extensions/drm-copilot run format
+EXIT_CODE: 0
+Output Summary: QA loop iteration 1 formatted `src/repo-automation-command-registration-admin.ts`, `src/repo-automation-command-registration-feature-workflows.ts`, `src/repo-automation-command-registration-types.ts`, and `src/repo-automation-command-registration.ts`; iteration 2 reran the same command and reported all matched files unchanged.
