+++ docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-26T19:56:29.2722493-04:00
+Command: npm --prefix extensions/drm-copilot run test:unit -- --coverage
+EXIT_CODE: 0
+Output Summary: Jest coverage passed on the clean QA pass with 32 suites and 349 tests. Coverage headlines: All files 95.51% lines, 95.51% statements, 87.07% branches, 95.87% functions; `repo-automation-command-registration-admin.ts` 97.09% lines, `repo-automation-command-registration-feature-workflows.ts` 100% lines, and `repo-automation-command-registration.ts` 100% lines.
