+++ docs/features/active/2026-04-26-codex-native-converter-164/feature-audit.2026-04-26T20-05.md
@@ -1,0 +1,31 @@
+# Feature Audit: codex-native-converter remediation closeout (#164)
+
+## Scope and Baseline
+
+This rerun reviews the residual remediation scope only: reducing `extensions/drm-copilot/src/repo-automation-command-registration.ts` to comply with the repository 500-line production-file limit while preserving the existing repo-automation command behavior. The review baseline is explicit base branch `development` at `0762f58a1451994999c2f49f2dbdc489120d138a`, with head `feature/codex-native-converter-164` at `b9542764a8271b83ecb075b7ca6edeb8575d1dfe`.
+
+## Acceptance Criteria Inventory
+
+- Source files: `docs/features/active/2026-04-26-codex-native-converter-164/spec.md` and `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`
+- User-story acceptance criteria remain delivered and already checked in `user-story.md`.
+- This remediation rerun adds the structural closeout conditions: the coordinator file must be `<= 500` lines, command behavior must remain unchanged, the TypeScript QA loop must pass, and refreshed review artifacts must be anchored to explicit base `development`.
+
+## Acceptance Criteria Evaluation
+
+| Item | Status | Evidence | Notes |
+| --- | --- | --- | --- |
+| Original user-story acceptance criteria remain satisfied | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md` | The remediation did not reopen feature behavior and all user-story criteria remain checked. |
+| `repo-automation-command-registration.ts` complies with the 500-line limit after the split | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/other/remediation-repo-automation-command-registration-lines-after.2026-04-26T19-48.md` | The coordinator file is now 23 lines. |
+| Existing command IDs, prompt flows, and activation wiring remain unchanged | PASS | `extensions/drm-copilot/src/extension.ts`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md` | Existing workflow tests covered the extracted command families and the clean Jest rerun passed. |
+| Final TypeScript QA loop passes with required coverage evidence | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-format.2026-04-26T19-48.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-lint.2026-04-26T19-48.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-typecheck.2026-04-26T19-48.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md` | The clean pass required two formatter iterations because the first run rewrote the new helper files. |
+| Review rerun basis is explicit and refreshed against `development` | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-pr-context-status.2026-04-26T19-48.md`, `artifacts/pr_context.summary.txt` | The rerun uses the explicit commit range `0762f58a1451994999c2f49f2dbdc489120d138a..b9542764a8271b83ecb075b7ca6edeb8575d1dfe`. |
+
+## Summary
+
+**Overall Feature Readiness:** PASS
+
+The residual remediation blocker is closed. The feature remains behaviorally complete, the structural policy violation has been removed, and the refreshed review evidence is sufficient for re-review.
+
+## Acceptance Criteria Check-off
+
+No additional acceptance-criteria checkboxes were changed during this remediation closeout because the authoritative user-story criteria were already checked and the spec checklist state did not require modification for this narrow structural fix.
