+++ docs/features/active/2026-04-26-codex-native-converter-164/feature-audit.2026-04-26T20-12.md
@@ -1,0 +1,63 @@
+# Feature Audit: codex-native-converter final post-remediation review (#164)
+
+## Scope and Baseline
+
+- **Base branch:** `development`
+- **Base commit:** `0762f58a1451994999c2f49f2dbdc489120d138a`
+- **Head commit reference:** `b9542764a8271b83ecb075b7ca6edeb8575d1dfe`
+- **Working-tree scope note:** The final review includes the current uncommitted repo-automation registration split that is visible in the workspace and reflected in `artifacts/pr_context.appendix.txt`.
+- **Requirements source:** `docs/features/active/2026-04-26-codex-native-converter-164/spec.md` and `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`
+- **Work mode:** `full-feature` from `issue.md`
+
+This final review verifies the full feature acceptance status after the two remediation loops. The original user-story acceptance criteria remain the authoritative delivery checklist, and the remediation closeout adds the requirement that the TypeScript wrapper surface remain structurally compliant without reopening delivered feature behavior.
+
+## Acceptance Criteria Inventory
+
+Authoritative acceptance-criteria source files:
+- `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`
+- `docs/features/active/2026-04-26-codex-native-converter-164/spec.md`
+
+Tracked feature acceptance criteria:
+1. Deterministic classification and concrete target-role assignment for supported GitHub Copilot and Claude source trees.
+2. Shared Python converter contract for the extension and MCP entry points.
+3. Explicit v1 support boundaries with explicit unsupported reporting.
+4. Output restricted to approved Codex-native surfaces, with repository prompt output gated by opt-in.
+5. Fail-closed hard-gate and handoff behavior.
+6. Semantic MCP rewrites to `drmCopilotExtension` when a safe mapping exists, with explicit reporting otherwise.
+7. Non-mutating review mode that emits the required artifact set.
+8. Apply mode requiring an explicit destination root and failing closed on unresolved requirements.
+9. Representative GitHub Copilot and Claude fixtures converting into reviewable outputs without flattening shared guidance.
+
+## Acceptance Criteria Evaluation
+
+| Item | Status | Evidence | Notes |
+| --- | --- | --- | --- |
+| 1. Deterministic classification and target-role assignment | PASS | `scripts/dev_tools/codex_native_converter/classifier.py`, `scripts/dev_tools/codex_native_converter/models.py`, `tests/scripts/dev_tools/codex_native_converter/test_classifier.py`, `tests/scripts/dev_tools/codex_native_converter/test_end_to_end.py` | The classifier and typed taxonomy remain the authoritative implementation and are covered by focused tests. |
+| 2. Extension and MCP entry points invoke the same Python converter contract | PASS | `extensions/drm-copilot/src/repo-automation-service.ts`, `extensions/drm-copilot/resources/templates/codex_native_converter.py`, `extensions/drm-copilot/test/repo-automation-service.codex-native-converter.test.ts` | The TypeScript layer remains a thin wrapper over the Python implementation. |
+| 3. Supported scope is explicit and unsupported items are reported explicitly | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/spec.md`, `scripts/dev_tools/codex_native_converter/inventory.py`, `scripts/dev_tools/codex_native_converter/validation.py` | The v1 support envelope is still explicit and fail-closed. |
+| 4. Outputs are limited to approved Codex-native surfaces | PASS | `scripts/dev_tools/codex_native_converter/mapping.py`, `scripts/dev_tools/codex_native_converter/validation.py`, `tests/scripts/dev_tools/codex_native_converter/test_mapping.py` | Mapping logic and tests preserve the approved target surface contract. |
+| 5. Hard-gate and handoff behavior remains fail-closed | PASS | `scripts/dev_tools/codex_native_converter/validation.py`, `tests/scripts/dev_tools/codex_native_converter/test_validation.py`, `tests/scripts/dev_tools/codex_native_converter/test_cli_entrypoints.py` | Blocking validation behavior remains explicit and non-discretionary. |
+| 6. Safe host-automation mappings rewrite to semantic MCP usage | PASS | `scripts/dev_tools/codex_native_converter/rewrites.py`, `scripts/dev_tools/codex_native_converter/validation.py`, `tests/scripts/dev_tools/codex_native_converter/test_validation.py` | Safe rewrites are preserved and unresolved mappings still fail closed. |
+| 7. Review mode is non-mutating and emits the required artifact set | PASS | `scripts/dev_tools/codex_native_converter/reporting.py`, `scripts/dev_tools/codex_native_converter/engine.py`, `tests/scripts/dev_tools/codex_native_converter/test_cli_review.py`, `tests/scripts/dev_tools/codex_native_converter/test_end_to_end.py` | Review mode still writes the required report set without destination-runtime mutation. |
+| 8. Apply mode requires an explicit destination root and fails closed when required | PASS | `scripts/dev_tools/codex_native_converter/cli.py`, `scripts/dev_tools/codex_native_converter/engine.py`, `tests/scripts/dev_tools/codex_native_converter/test_cli_apply.py` | Apply-mode guardrails remain intact. |
+| 9. Representative GitHub Copilot and Claude fixtures convert successfully into reviewable outputs | PASS | `tests/fixtures/codex_native_converter/**`, `tests/scripts/dev_tools/codex_native_converter/test_end_to_end.py` | Both representative ecosystems remain covered by the checked-in fixture tests. |
+| Remediation closeout: TypeScript wrapper surface is structurally compliant in the current working tree | PASS | `artifacts/pr_context.appendix.txt`, direct working-tree inspection of `extensions/drm-copilot/src/repo-automation-command-registration*.ts`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md` | The coordinator and helper modules are now under the 500-line limit and the clean remediation QA pass remains valid for the split surface. |
+
+## Summary
+
+**Overall Feature Readiness:** PASS
+
+The feature acceptance criteria remain satisfied, the Python-first architecture is intact, and the final remediation loop closed the remaining TypeScript structural blocker. The current branch state is ready for merge review and does not require another remediation loop.
+
+## Acceptance Criteria Check-off
+
+The authoritative `user-story.md` acceptance-criteria checkboxes were already checked before this final review, and the current review found no reason to reopen them.
+
+### Acceptance Criteria Status
+- Source: `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`, `docs/features/active/2026-04-26-codex-native-converter-164/spec.md`
+- Total AC items: 9
+- Checked off (delivered): 9
+- Remaining (unchecked): 0
+- Items remaining: None.
+
+No source-file checkbox changes were required during this final review.
