+++ docs/features/active/2026-04-26-codex-native-converter-164/feature-audit.2026-04-26T23-41.md
@@ -16,15 +16,16 @@ This rerun reviews the remediation for findings R-1 through R-3 only. Baseline l
 | --- | --- | --- | --- |
 | Original user-story acceptance criteria remain satisfied | PASS | `docs/features/active/2026-04-26-codex-native-converter-164/user-story.md`, prior `feature-audit.2026-04-26T19-20.md` | The remediation did not reopen feature behavior. |
 | `extension.ts` and `repo-automation-service.ts` comply with the 500-line limit | PASS | `remediation-extension-lines-after.2026-04-26T19-20.md`, `remediation-repo-automation-service-lines-after.2026-04-26T19-20.md` | Both touched production files are now below 500 lines. |
+| No new structural blocker is introduced by the split | FAIL | `remediation-repo-automation-command-registration-lines-after.2026-04-26T19-20.md` | `repo-automation-command-registration.ts` is 513 lines and still exceeds the repository production-file limit. |
 | Final TypeScript QA loop passes with required coverage evidence | PASS | `remediation-typescript-format.2026-04-26T19-20.md`, `remediation-typescript-lint.2026-04-26T19-20.md`, `remediation-typescript-typecheck.2026-04-26T19-20.md`, `remediation-typescript-test-coverage.2026-04-26T19-20.md` | Final clean pass completed after the required loop restarts. |
 | Review rerun basis is explicit | PASS | `remediation-pr-context-status.2026-04-26T19-20.md`, `artifacts/pr_context.summary.txt` | The rerun now uses a non-empty commit range against `development`. |
 
 ## Summary
 
-**Overall Feature Readiness:** PASS
+**Overall Feature Readiness:** PARTIAL
 
-The remediation closed the only remaining review blocker without reopening delivered feature behavior. The branch is structurally compliant in the remediated scope and is ready for normal review follow-up.
+The remediation closed the two original blockers without reopening delivered feature behavior, but it left one new structural blocker in the extracted command-registration module. The branch needs one additional split before the remediation can be considered fully complete.
 
 ## Acceptance Criteria Check-off
 
-No additional acceptance-criteria checkboxes were changed during remediation because the original feature criteria were already PASS before the structural fix. The remediation preserved that PASS state while closing the policy blocker.
+No additional acceptance-criteria checkboxes were changed during remediation because the original feature criteria were already PASS before the structural fix. The remediation preserved delivered behavior, but the remaining structural blocker prevents a final all-clear rerun.
