+++ docs/features/active/2026-04-26-codex-native-converter-164/policy-audit.2026-04-26T20-05.md
@@ -1,0 +1,87 @@
+# Policy Audit: codex-native-converter remediation closeout (#164)
+
+**Reviewer:** atomic_executor (GitHub Copilot)
+**Review scope:** Post-remediation branch state for `feature/codex-native-converter-164` against explicit base `development`
+**Base branch:** `development`
+**Base commit:** `0762f58a1451994999c2f49f2dbdc489120d138a`
+**Head branch:** `feature/codex-native-converter-164`
+**Head commit:** `b9542764a8271b83ecb075b7ca6edeb8575d1dfe`
+**Code Under Test:** `extensions/drm-copilot/src/repo-automation-command-registration.ts`, `extensions/drm-copilot/src/repo-automation-command-registration-admin.ts`, `extensions/drm-copilot/src/repo-automation-command-registration-feature-workflows.ts`, `extensions/drm-copilot/src/repo-automation-command-registration-types.ts`, `extensions/drm-copilot/src/extension.ts`, and remediation evidence under `docs/features/active/2026-04-26-codex-native-converter-164/evidence/`
+
+## Executive Summary
+
+This remediation rerun closes the residual structural blocker from issue #164. The previously oversized `repo-automation-command-registration.ts` file is now a 23-line coordinator that assembles two focused helper modules: one for admin or repository-support registrations and one for feature-promotion workflow registrations. The public `registerRepoAutomationCommands` export remains unchanged in `extension.ts`, and the TypeScript QA loop completed with a clean final pass after one formatter-triggered restart.
+
+## 1. General Unit Test Policy Compliance
+
+- TypeScript baseline coverage artifact: `docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-2-typescript-test-coverage.2026-04-26T19-48.md`
+- TypeScript post-change coverage artifact: `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md`
+- PowerShell baseline coverage artifact: N/A - remediation touched no PowerShell files.
+- PowerShell post-change coverage artifact: N/A - remediation touched no PowerShell files.
+- Per-language comparison summary: TypeScript baseline 95.49% lines, post-change 95.51% lines, new or changed production files 97.09% to 100% lines.
+
+### 1.2.1 Per-Language Coverage Comparison
+
+- TypeScript: Baseline: 95.49% lines; Post-change: 95.51% lines; Change: +0.02%; New/changed-code coverage: `repo-automation-command-registration-admin.ts` 97.09%, `repo-automation-command-registration-feature-workflows.ts` 100%, `repo-automation-command-registration.ts` 100%; Disposition: PASS; Evidence: `docs/features/active/2026-04-26-codex-native-converter-164/evidence/baseline/remediation-2-typescript-test-coverage.2026-04-26T19-48.md`, `docs/features/active/2026-04-26-codex-native-converter-164/evidence/qa-gates/remediation-2-typescript-test-coverage.2026-04-26T19-48.md`
+
+## 2. General Code Change Policy Compliance
+
+The remediation remained limited to the residual TypeScript structural blocker identified in `remediation-inputs.2026-04-26T19-48.md`. The command-registration surface now follows the repository module-size rule: `repo-automation-command-registration.ts` is 23 lines, `repo-automation-command-registration-admin.ts` is focused on admin and review-support commands, and `repo-automation-command-registration-feature-workflows.ts` is focused on feature-entry and promotion workflows.
+
+## 3. Language-Specific Code Change Policy Compliance
+
+The TypeScript changes remain within the extension package, preserve typed service delegation, and keep `extension.ts` as a thin activation coordinator. No suppressions or API-breaking changes were introduced.
+
+## 4. Language-Specific Unit Test Policy Compliance
+
+Existing focused Jest workflow tests already exercised the extracted command families, and the clean package-wide Jest coverage rerun passed 32 suites and 349 tests. All newly created production files are at or above the repository 90% coverage threshold.
+
+## 5. Test Coverage Detail
+
+| Language | Files Changed | Tests | Test Result | Baseline Coverage | Post-Change Coverage | New Code Coverage |
+| --- | --- | --- | --- | --- | --- | --- |
+| TypeScript | `repo-automation-command-registration.ts`, `repo-automation-command-registration-admin.ts`, `repo-automation-command-registration-feature-workflows.ts`, `repo-automation-command-registration-types.ts` | Jest unit tests with coverage | PASS | 95.49% lines | 95.51% lines | 97.09% to 100% lines |
+
+## 6. Test Execution Metrics
+
+- Formatter: `npm --prefix extensions/drm-copilot run format` → PASS on iteration 2 after iteration 1 formatted the new registration modules.
+- Linter: `npm --prefix extensions/drm-copilot run lint` → PASS.
+- Type checker: `npm --prefix extensions/drm-copilot run typecheck` → PASS.
+- Test runner: `npm --prefix extensions/drm-copilot run test:unit -- --coverage` → PASS, 32 suites, 349 tests.
+- PR context refresh: `poetry run python -m scripts.dev_tools.pr_context.collector --base development` → PASS with commit-range review `0762f58a1451994999c2f49f2dbdc489120d138a..b9542764a8271b83ecb075b7ca6edeb8575d1dfe`.
+
+## 7. Code Quality Checks
+
+- Residual blocker closed: `extensions/drm-copilot/src/repo-automation-command-registration.ts` is now 23 lines.
+- New helper modules are focused by command family and remain below the 500-line production-file limit.
+- `extensions/drm-copilot/src/extension.ts` still imports and invokes the same public `registerRepoAutomationCommands` helper.
+- PR context is anchored to explicit base `development` and current head `feature/codex-native-converter-164`.
+
+## 8. Gaps and Exceptions
+
+No remaining structural blocker was identified in the remediation scope. The TypeScript QA loop required two formatter iterations because the first formatter pass rewrote the new helper files; the second pass was clean and serves as the final authoritative pass.
+
+## 9. Summary of Changes
+
+The remediation split the command-registration logic into three focused modules, preserved the existing command IDs and prompt flows, recorded refreshed baseline and QA evidence, and reran the `development`-based review artifacts against the current branch state.
+
+## 10. Compliance Verdict
+
+PASS. The residual `repo-automation-command-registration.ts` file-size blocker is closed, the clean TypeScript QA pass succeeded, and the refreshed review basis is explicit.
+
+## Appendix A: Test Inventory
+
+- `extensions/drm-copilot/test/extension.workflow-commands.test.ts`
+- `extensions/drm-copilot/test/extension.collect-pr-context.test.ts`
+- `extensions/drm-copilot/test/extension.new-active-feature-folder.test.ts`
+- `extensions/drm-copilot/test/extension.potential-to-issue.test.ts`
+- `extensions/drm-copilot/test/repo-automation-dispatch.test.ts`
+- Package-wide Jest coverage run via `npm --prefix extensions/drm-copilot run test:unit -- --coverage`
+
+## Appendix B: Toolchain Commands Reference
+
+- `npm --prefix extensions/drm-copilot run format`
+- `npm --prefix extensions/drm-copilot run lint`
+- `npm --prefix extensions/drm-copilot run typecheck`
+- `npm --prefix extensions/drm-copilot run test:unit -- --coverage`
+- `poetry run python -m scripts.dev_tools.pr_context.collector --base development`
