+++ extensions/drm-copilot/src/repo-automation-command-registration-types.ts
@@ -1,0 +1,9 @@
+import * as vscode from "vscode";
+
+import type { RepoAutomationService } from "./repo-automation-service";
+
+export interface RepoAutomationCommandRegistrationOptions {
+  readonly context: vscode.ExtensionContext;
+  readonly output: vscode.OutputChannel;
+  readonly service: RepoAutomationService;
+}
